Face/Code/requirements.txt
flask==3.1.1
flask-cors==6.0.1
face-recognition==1.3.0
opencv-python-headless==4.11.0.86
torch==2.5.1
numpy==1.26.4
pillow==11.2.1
dlib-bin==19.24.6
easydict==1.10
tensorboardX==2.5.1
